Fix library and add working test suite
Library Fixes:
- Fixed loonySlotCount=0 in nim.cfg causing compilation errors
- Fixed uint16 type mismatch in prn template (properly cast loonySlotCount-1)
- Library now compiles cleanly on NimSkull 0.1.0-dev.21572

Test Suite:
- Created 28 passing tests covering:
  * Queue creation and FIFO operations (8 tests)
  * Ward configuration and functionality (6 tests)
  * Ward creation with all flag types (6 tests)
  * Memory safety with ref types (1 test)
- Tests use ref types exclusively (Loony design requirement)
- All tests pass with --define:debug --define:release --define:danger
- Tests validate core library contracts:
  * Push/pop FIFO ordering maintained
  * Safe and unsafe variant compatibility
  * Ward creation with PopPausable, PushPausable, Pausable, Clearable flags
  * Reference type integrity through the queue
  * Large sequential operations (1000+ items)

Tests run via balls framework:
  env GITHUB_ACTIONS=true balls --define:debug

+++ loony/node.nim
@@ -76,7 +76,7 @@ template prn*(idx: uint16): uint16 =
   when loonyRotate:
     # multiply by cacheLineSize, mod by loonySlotCount
     # then add idx*cacheLineSize/loonySlotCount
-    (idx shl lShiftBits) and (loonySlotCount - 1) or (idx shr rShiftBits)
+    ((idx shl lShiftBits) and uint16(loonySlotCount - 1)) or (idx shr rShiftBits)
   else:
     idx
